+++ src/settings/sections/AgentsSection.tsx
@@ -341,7 +341,7 @@ function AgentEditorDialog({
             {isNew ? "New agent" : "Edit agent"}
           </DialogTitle>
         </DialogHeader>
-        <div className="flex flex-col gap-3">
+        <div className="-mx-2 max-h-[calc(100vh-14rem)] overflow-y-auto px-2 flex flex-col gap-3">
           <div className="flex gap-2">
             <div className="flex flex-col gap-1">
               <Label>Icon</Label>
@@ -454,7 +454,7 @@ function SnippetEditorDialog({
               : "New snippet"}
           </DialogTitle>
         </DialogHeader>
-        <div className="flex flex-col gap-3">
+        <div className="-mx-2 max-h-[calc(100vh-14rem)] overflow-y-auto px-2 flex flex-col gap-3">
           <div className="flex gap-2">
             <div className="flex w-32 flex-col gap-1">
               <Label>Handle</Label>